refactor(auth): modular session store with config factory
