+++ packages/auth/package.json
@@ -13,7 +13,8 @@
     "iron-session": "^6.3.1",
     "otplib": "^12.0.1",
     "@acme/platform-core": "workspace:*",
-    "@upstash/redis": "^1.35.3"
+    "@upstash/redis": "^1.35.3",
+    "@acme/config": "workspace:*"
   },
   "devDependencies": {
     "next": "^15.3.4"
