+++ packages/auth/src/redisStore.ts
@@ -1,0 +1,45 @@
+import type { Redis } from "@upstash/redis";
+import type { SessionRecord, SessionStore } from "./store";
+
+export class RedisSessionStore implements SessionStore {
+  constructor(private client: Redis, private ttl: number) {}
+
+  private key(id: string) {
+    return `session:${id}`;
+  }
+
+  private customerKey(customerId: string) {
+    return `customer_sessions:${customerId}`;
+  }
+
+  async get(id: string): Promise<SessionRecord | null> {
+    const data = await this.client.get<Record<string, any>>(this.key(id));
+    if (!data) return null;
+    return { ...data, createdAt: new Date(data.createdAt) } as SessionRecord;
+  }
+
+  async set(record: SessionRecord): Promise<void> {
+    await this.client.set(this.key(record.sessionId), record, { ex: this.ttl });
+    await this.client.sadd(this.customerKey(record.customerId), record.sessionId);
+    await this.client.expire(this.customerKey(record.customerId), this.ttl);
+  }
+
+  async delete(id: string): Promise<void> {
+    const rec = await this.get(id);
+    await this.client.del(this.key(id));
+    if (rec) {
+      await this.client.srem(this.customerKey(rec.customerId), id);
+    }
+  }
+
+  async list(customerId: string): Promise<SessionRecord[]> {
+    const ids = await this.client.smembers<string>(this.customerKey(customerId));
+    if (!ids || ids.length === 0) return [];
+    const records = await this.client.mget<Record<string, any>>(
+      ...ids.map((id) => this.key(id))
+    );
+    return records
+      .filter((r): r is Record<string, any> => r !== null)
+      .map((r) => ({ ...r, createdAt: new Date(r.createdAt) } as SessionRecord));
+  }
+}
